+++ src/main/resources/application.properties
@@ -9,6 +9,11 @@ spring.flyway.locations=classpath:db/migration,classpath:db/testdata
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL57Dialect
 
+# SECURITY
+spring.security.oauth2.resourceserver.opaquetoken.introspection-uri=http://localhost:8081/oauth/check_token
+spring.security.oauth2.resourceserver.opaquetoken.client-id=algafood-web
+spring.security.oauth2.resourceserver.opaquetoken.client-secret=web123
+
 # JACKSON
 spring.jackson.deserialization.fail-on-unknown-properties=true
 spring.jackson.deserialization.fail-on-ignored-properties=true
@@ -46,4 +51,4 @@ spring.freemarker.settings.locale=pt_BR
 
 # LOG
 logging.level.org.springframework.boot=INFO
-logging.level.org.hibernate=ERROR]
+logging.level.org.hibernate=ERROR
